fix: maven 版本信息读取pom文件信息 #193

+++ src/backend/maven/biz-maven/src/main/kotlin/com/tencent/bkrepo/maven/pojo/Basic.kt
@@ -8,7 +8,7 @@ data class Basic(
     val fullPath: String,
     val lastModifiedBy: String,
     val lastModifiedDate: String,
-    val downloadCount: Int,
+    val downloadCount: Long,
     val sha256: String?,
     val md5: String?,
     val stageTag: List<String>?,
